+++ lib/hipcall_whichtech.ex
@@ -3,16 +3,13 @@ defmodule HipcallWhichtech do
   Documentation for `HipcallWhichtech`.
   """
 
-  @doc """
-  Hello world.
 
-  ## Examples
-
-      iex> HipcallWhichtech.hello()
-      :world
-
-  """
-  def hello do
-    :world
+  def html(url) do
+    case Finch.build(:get, url) |> Finch.request(HipcallWhichtechFinch) do
+      {:ok, %Finch.Response{status: 200, body: body, headers: _headers, trailers: _trailers}} ->
+        {:ok, body}
+      {:error, exception} ->
+        {:error, exception}
+    end
   end
 end