+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## v0.1.0 (2024-01-28)
 
+- Add `Wordpress` detector
+- Add `Hubspot` detector
+  Add `Shopify` detector
 - Add CHANGELOG.md
 - Add `models` endpoint
 - Add `Config` overwrite feature